Split `file_time.rs` into submodules
